+++ sm4sh_blender/__init__.py
@@ -5,7 +5,7 @@ bl_info = {
     "description": "Import and export Smash 4 models and animations",
     "author": "ScanMountGoat (SMG)",
     "version": (0, 3, 1),
-    "blender": (4, 1, 0),
+    "blender": (5, 1, 0),
     "location": "File > Import-Export",
     "warning": "",
     "doc_url": "https://github.com/ScanMountGoat/sm4sh_blender/wiki",