+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## unreleased
+### Fixed
+* Fixed an issue where default material creation would fail whenn exporting meshes with no material assigned.
+
 ## 0.3.1 - 2025-12-08
 ### Fixed
 * Fixed an issue where exported nud models would not have correctly normalized vertex weights in some cases.