+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 * Changed nud export to report a warning if a mesh does not have all attributes required by the shaders.
 
 ### Fixed
-* Fixed an issue where default material creation would fail whenn exporting meshes with no material assigned.
+* Fixed an issue where default material creation would fail when exporting meshes with no material assigned.
 
 ## 0.3.1 - 2025-12-08
 ### Fixed